refactored tabs to be a bit more reusable
