+++ src/frontend/components/split-pane/split-pane.html
@@ -4,7 +4,7 @@
    class="flex flex-auto">
   <div
      #primary
-     class="col overflow-hidden flex maxheight-100pct">
+     class="col overflow-hidden flex maxheight-100pct flex-column">
     <div class="status-bar minwidth-100pct" style="max-height: 30px">
       Zone Busy Time: {{zoneBusyTime ? zoneBusyTime / 1000 : 'N/A'}}s
     </div>
